+++ src/main/java/project/tripplan/domain/user/repository/UserRepositoryCustom.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import project.tripplan.domain.user.dto.UserBookmarkRes;
 import project.tripplan.domain.user.dto.UserPlanRes;
 import project.tripplan.domain.user.entity.User;
 
@@ -14,5 +13,4 @@ public interface UserRepositoryCustom {
 
 	Page<UserPlanRes> findPlansByUserId(Long userId, Pageable pageable);
 
-	Page<UserBookmarkRes> findBookmarksByUserId(Long userId, Pageable pageable);
 }
