+++ src/main/java/project/tripplan/domain/user/repository/UserRepositoryCustomImpl.java
@@ -13,10 +13,8 @@ import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
-import project.tripplan.domain.bookmark.entity.QBookmark;
 import project.tripplan.domain.plan.entity.QPlan;
 import project.tripplan.domain.plan.entity.QPlanPlaceCategory;
-import project.tripplan.domain.user.dto.UserBookmarkRes;
 import project.tripplan.domain.user.dto.UserPlanRes;
 import project.tripplan.domain.user.entity.QUser;
 import project.tripplan.domain.user.entity.User;
@@ -27,7 +25,6 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
 	private final JPAQueryFactory qf;
 	private final QUser user = QUser.user;
 	private final QPlan plan = QPlan.plan;
-	private final QBookmark bookmark = QBookmark.bookmark;
 	private final QPlanPlaceCategory placeCategory = QPlanPlaceCategory.planPlaceCategory;
 
 	@Override
@@ -72,36 +69,4 @@ public class UserRepositoryCustomImpl implements UserRepositoryCustom {
 		return new PageImpl<>(content, pageable, total);
 	}
 
-	@Override
-	public Page<UserBookmarkRes> findBookmarksByUserId(Long userId, Pageable pageable) {
-		List<UserBookmarkRes> content = qf
-			.select(Projections.constructor(
-				UserBookmarkRes.class,
-				bookmark.id,
-				plan.id,
-				plan.title,
-				plan.imageUrl.as("thumbnail"),
-				Expressions.stringTemplate(
-					"group_concat(DISTINCT {0})",
-					placeCategory.placeCategory.name
-				)
-			))
-			.from(bookmark)
-			.join(bookmark.plan, plan)
-			.leftJoin(plan.planPlaceCategories, placeCategory)
-			.where(bookmark.user.id.eq(userId))
-			.groupBy(bookmark.id, plan.id)
-			.orderBy(bookmark.createdAt.desc())
-			.offset(pageable.getOffset())
-			.limit(pageable.getPageSize())
-			.fetch();
-
-		long total = qf
-			.select(plan.count())
-			.from(plan)
-			.where(plan.user.id.eq(userId))
-			.fetchOne();
-
-		return new PageImpl<>(content, pageable, total);
-	}
 }
